+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/AdMessageCodec.java
@@ -36,6 +36,7 @@ class AdMessageCodec extends StandardMessageCodec {
   // The type values below must be consistent for each platform.
   private static final byte VALUE_AD_SIZE = (byte) 128;
   private static final byte VALUE_AD_REQUEST = (byte) 129;
+  private static final byte VALUE_FLUID_AD_SIZE = (byte) 130;
   private static final byte VALUE_REWARD_ITEM = (byte) 132;
   private static final byte VALUE_LOAD_AD_ERROR = (byte) 133;
   private static final byte VALUE_ADMANAGER_AD_REQUEST = (byte) 134;
@@ -52,7 +53,7 @@ class AdMessageCodec extends StandardMessageCodec {
   static final byte VALUE_VIDEO_OPTIONS = (byte) 145;
   static final byte VALUE_LOCATION_PARAMS = (byte) 146;
 
-  @NonNull final Context context;
+  @NonNull Context context;
   @NonNull final FlutterAdSize.AdSizeFactory adSizeFactory;
 
   AdMessageCodec(@NonNull Context context) {
@@ -64,6 +65,10 @@ class AdMessageCodec extends StandardMessageCodec {
   AdMessageCodec(@NonNull Context context, @NonNull FlutterAdSize.AdSizeFactory adSizeFactory) {
     this.context = context;
     this.adSizeFactory = adSizeFactory;
+  }
+
+  void setContext(@NonNull Context context) {
+    this.context = context;
   }
 
   @Override
@@ -193,6 +198,8 @@ class AdMessageCodec extends StandardMessageCodec {
         return new FlutterAdSize(
             (Integer) readValueOfType(buffer.get(), buffer),
             (Integer) readValueOfType(buffer.get(), buffer));
+      case VALUE_FLUID_AD_SIZE:
+        return new FlutterAdSize.FluidAdSize();
       case VALUE_AD_REQUEST:
         return new FlutterAdRequest.Builder()
             .setKeywords((List<String>) readValueOfType(buffer.get(), buffer))
@@ -299,6 +306,8 @@ class AdMessageCodec extends StandardMessageCodec {
       writeValue(stream, size.width);
     } else if (value instanceof FlutterAdSize.SmartBannerAdSize) {
       stream.write(VALUE_SMART_BANNER_AD_SIZE);
+    } else if (value instanceof FlutterAdSize.FluidAdSize) {
+      stream.write(VALUE_FLUID_AD_SIZE);
     } else {
       stream.write(VALUE_AD_SIZE);
       writeValue(stream, value.width);